Correccion migration Users

+++ database/migrations/2020_04_01_107956_create_usuarios_table.php
@@ -20,7 +20,7 @@ class CreateUsuariosTable extends Migration
             $table->string('last_name');
             $table->string('email');
             $table->string('password');
-            $table->bigInteger('rol_id')->unsigned();
+            $table->integer('rol_id')->unsigned();
             $table->timestamps();
             
             $table->foreign('rol_id')->references('id')->on('roles');
